Move python finder to separate file
Python will be also used by build stage
